refactor: 3단계 선택지 추림·완료 처리를 화면에서 뺌 #88
invite-setup 297 → 169. 무엇을 고를 수 있는지(invite-choices)와 [완료]가 하는 일
(use-invite-commit), 아래 버튼 줄(invite-footer)을 갈랐다.
